components: delete datetimepicker component
